move python wrapper from moe.py to mhc.py Manifold-Constrained Hyper-Connections
